refactor app.js/start
